Thêm chức năng tìm kiếm theo tiêu chí

+++ src/product/controller/Serializator.java
@@ -1,22 +1,54 @@
 package product.controller;
 import java.io.*;
+import java.nio.file.FileAlreadyExistsException;
 
 public class Serializator {
-    public ProductManager readFromFile(String pathSource) throws IOException, ClassNotFoundException {
-        ObjectInputStream object = new ObjectInputStream(new FileInputStream(pathSource));
-        ProductManager productManager = (ProductManager) object.readObject();
-        object.close();
+    private ProductManager productManager;
+    private File file;
+
+    public Serializator(ProductManager productManager) {
+        this.productManager = productManager;
+    }
+
+    public Serializator() {
+    }
+
+    public ProductManager getProductManager() {
         return productManager;
     }
 
-    public void writeToFile(String pathDest, ProductManager productManager) throws IOException, ClassCastException {
+    public File getFile() throws IOException {
+        if (file == null) {
+            throw new FileNotFoundException("File not found");
+        }
+        return file;
+    }
+
+    public ProductManager readFromFile(String pathSource) throws IOException, ClassNotFoundException {
+        ObjectInputStream object = new ObjectInputStream(new FileInputStream(pathSource));
+        productManager = (ProductManager) object.readObject();
+        object.close();
+        file = new File(pathSource);
+        return productManager;
+    }
+
+    public void writeToFile(String pathDest) throws IOException, ClassCastException {
+        file = new File(pathDest);
+        if (file.exists()) {
+            throw new FileAlreadyExistsException(pathDest);
+        }
         ObjectOutputStream object = new ObjectOutputStream(new FileOutputStream(pathDest));
         object.writeObject(productManager);
         object.close();
     }
 
+    public void writeToFile() throws IOException, ClassCastException {
+        ObjectOutputStream object = new ObjectOutputStream(new FileOutputStream(file.getAbsolutePath()));
+        object.writeObject(productManager);
+        object.close();
+    }
+
     public void exportToText(String pathDest, Object object) throws IOException {
-        Writer out;
         BufferedWriter writer = new BufferedWriter(new FileWriter(pathDest));
         writer.write(object.toString());
         writer.close();
